Переменные. [#172635 #334332]

+++ src/main/java/ru/job4j/Variables.java
@@ -4,7 +4,7 @@ public class Variables {
     public static void main(String[] args) {
         int age = 18;
         System.out.println(age);
-        double length = 78.0332;
+        float length = 78.0332f;
         System.out.println(length);
         boolean baby = true;
         System.out.println(baby);
